[fix] age ratio

+++ src/main/java/mutsa/backend/BusinessDistrict/dto/ppl/BusinessAge.java
@@ -2,10 +2,10 @@ package mutsa.backend.BusinessDistrict.dto.ppl;
 
 public record BusinessAge(
         String dong,
-        Long pplAge10,
-        Long pplAge20,
-        Long pplAge30,
-        Long pplAge40,
-        Long pplAge50,
-        Long pplAge60
+        double pplAge10,
+        double pplAge20,
+        double pplAge30,
+        double pplAge40,
+        double pplAge50,
+        double pplAge60
 ) {}
